+++ test/test_orm.py
@@ -12,32 +12,32 @@ def test_1():
     >>> db.metadata.drop_all()
     >>> class Test(Model):
     ...     username = Field(unicode)
-    ...     year = Field(int, default=0)
+    ...     year = Field(int, default=30)
     ...     birth = Field(datetime.date)
     >>> a = Test(username='limodou', birth='2011-03-04')
     >>> a.save()
     True
     >>> a
-    <Test {'username':u'limodou','year':0,'birth':datetime.date(2011, 3, 4),'id':1}>
+    <Test {'username':u'limodou','year':30,'birth':datetime.date(2011, 3, 4),'id':1}>
     >>> b = Test(username=u'limodou1')
     >>> b.save()
     True
     >>> b
-    <Test {'username':u'limodou1','year':0,'birth':None,'id':2}>
+    <Test {'username':u'limodou1','year':30,'birth':None,'id':2}>
     >>> print list(Test.all())
-    [<Test {'username':u'limodou','year':0,'birth':datetime.date(2011, 3, 4),'id':1}>, <Test {'username':u'limodou1','year':0,'birth':None,'id':2}>]
+    [<Test {'username':u'limodou','year':30,'birth':datetime.date(2011, 3, 4),'id':1}>, <Test {'username':u'limodou1','year':30,'birth':None,'id':2}>]
     >>> print Test.count()
     2
     >>> a.username
     u'limodou'
     >>> list(Test.filter(Test.c.username==u'limodou'))
-    [<Test {'username':u'limodou','year':0,'birth':datetime.date(2011, 3, 4),'id':1}>]
+    [<Test {'username':u'limodou','year':30,'birth':datetime.date(2011, 3, 4),'id':1}>]
     >>> c = Test.get(1)
     >>> c
-    <Test {'username':u'limodou','year':0,'birth':datetime.date(2011, 3, 4),'id':1}>
+    <Test {'username':u'limodou','year':30,'birth':datetime.date(2011, 3, 4),'id':1}>
     >>> c = Test.get(Test.c.id==1)
     >>> c
-    <Test {'username':u'limodou','year':0,'birth':datetime.date(2011, 3, 4),'id':1}>
+    <Test {'username':u'limodou','year':30,'birth':datetime.date(2011, 3, 4),'id':1}>
     >>> Test.remove(1)
     >>> Test.count()
     1